no fast movement for striped test. better help on failures.
we have now found an artifact failure for striped mode on a Dell
2001FP

manifest only when help text is shown - at least seems that way
and the artifacts sometimes only manifest after the sprite position is
reset

some form of complex full screen and/or stateful processing is taking
place. this is undesirable and constitues a fail.